requirements.txt
torch
torchvision
torchaudio
git+https://github.com/openai/shap-e.git
Pillow
gradio
